+++ zmeventnotification/zmeventnotification/zmes_hook_helpers/utils.py
@@ -25,6 +25,107 @@ from urllib.error import HTTPError
 #resize polygons based on analysis scale
 
 
+
+
+def convert_config_to_ml_sequence():
+    ml_options={}
+    for ds in g.config['detection_sequence']:
+        if ds == 'object':
+        
+            ml_options['object'] = {
+                'general':{
+                    'pattern': g.config['object_detection_pattern'],
+                    'disable_locks': g.config['disable_locks'],
+                    'same_model_sequence_strategy': 'first' # 'first' 'most', 'most_unique'
+
+                },
+                'sequence': [{
+                    'tpu_max_processes': g.config['tpu_max_processes'],
+                    'tpu_max_lock_wait': g.config['tpu_max_lock_wait'],
+                    'gpu_max_processes': g.config['gpu_max_processes'],
+                    'gpu_max_lock_wait': g.config['gpu_max_lock_wait'],
+                    'cpu_max_processes': g.config['cpu_max_processes'],
+                    'cpu_max_lock_wait': g.config['cpu_max_lock_wait'],
+                    'max_detection_size': g.config['max_detection_size'],
+                    'object_config':g.config['object_config'],
+                    'object_weights':g.config['object_weights'],
+                    'object_labels': g.config['object_labels'],
+                    'object_min_confidence': g.config['object_min_confidence'],
+                    'object_framework':g.config['object_framework'],
+                    'object_processor': g.config['object_processor'],
+                }]
+            }
+        elif ds == 'face':
+            ml_options['face'] = {
+                'general':{
+                    'pattern': g.config['face_detection_pattern'],
+                    'same_model_sequence_strategy': 'first',
+                #    'pre_existing_labels':['person'],
+                },
+                'sequence': [{
+                    'tpu_max_processes': g.config['tpu_max_processes'],
+                    'tpu_max_lock_wait': g.config['tpu_max_lock_wait'],
+                    'gpu_max_processes': g.config['gpu_max_processes'],
+                    'gpu_max_lock_wait': g.config['gpu_max_lock_wait'],
+                    'cpu_max_processes': g.config['cpu_max_processes'],
+                    'cpu_max_lock_wait': g.config['cpu_max_lock_wait'],
+                    'face_detection_framework': g.config['face_detection_framework'],
+                    'face_recognition_framework': g.config['face_recognition_framework'],
+                    'face_processor': g.config['face_processor'],
+                    'known_images_path': g.config['known_images_path'],
+                    'face_model': g.config['face_model'],
+                    'face_train_model':g.config['face_train_model'],
+                    'unknown_images_path': g.config['unknown_images_path'],
+                    'unknown_face_name': g.config['unknown_face_name'],
+                    'save_unknown_faces': g.config['save_unknown_faces'],
+                    'save_unknown_faces_leeway_pixels': g.config['save_unknown_faces_leeway_pixels'],
+                    'face_recog_dist_threshold': g.config['face_recog_dist_threshold'],
+                    'face_num_jitters': g.config['face_num_jitters'],
+                    'face_upsample_times':g.config['face_upsample_times']
+                }]
+
+            }
+        elif ds == 'alpr':
+            ml_options['alpr'] = {
+                'general':{
+                    'pattern': g.config['alpr_detection_pattern'],
+                    'same_model_sequence_strategy': 'first',
+                #    'pre_existing_labels':['person'],
+                },
+                'sequence': [{
+                    'tpu_max_processes': g.config['tpu_max_processes'],
+                    'tpu_max_lock_wait': g.config['tpu_max_lock_wait'],
+                    'gpu_max_processes': g.config['gpu_max_processes'],
+                    'gpu_max_lock_wait': g.config['gpu_max_lock_wait'],
+                    'cpu_max_processes': g.config['cpu_max_processes'],
+                    'cpu_max_lock_wait': g.config['cpu_max_lock_wait'],
+                    'alpr_service': g.config['alpr_service'],
+                    'alpr_url': g.config['alpr_url'],
+                    'alpr_key': g.config['alpr_key'],
+                    'alpr_api_type': g.config['alpr_api_type'],
+                    'platerec_stats': g.config['platerec_stats'],
+                    'platerec_regions': g.config['platerec_regions'],
+                    'platerec_min_dscore': g.config['platerec_min_dscore'],
+                    'platerec_min_score': g.config['platerec_min_score'],
+                    'openalpr_recognize_vehicle': g.config['openalpr_recognize_vehicle'],
+                    'openalpr_country': g.config['openalpr_country'],
+                    'openalpr_state': g.config['openalpr_state'],
+                    'openalpr_min_confidence': g.config['openalpr_min_confidence'],
+                    'openalpr_cmdline_binary': g.config['openalpr_cmdline_binary'],
+                    'openalpr_cmdline_params': g.config['openalpr_cmdline_params'],
+                    'openalpr_cmdline_min_confidence': g.config['openalpr_cmdline_min_confidence'],
+                }]
+
+            }
+    ml_options['general'] =   {
+            'model_sequence': ','.join(str(e) for e in g.config['detection_sequence'])
+            #'model_sequence': 'object,face',        
+    }
+    if g.config['detection_mode'] == 'all':
+        g.logger.Debug(3, 'Changing detection_mode from all to most_models to adapt to new features')
+        g.config['detection_mode'] = 'most_models'
+    return ml_options
+
 def rescale_polygons(xfactor, yfactor):
     newps = []
     for p in g.polygons:
@@ -112,6 +213,10 @@ def import_zm_zones(mid, reason):
     #g.logger.Debug(1,'Found motion zones provided in alarm cause: {}'.format(reason_zones))
 
     for item in j['zones']:
+        #print ('********* ITEM TYPE {}'.format(item['Zone']['Type']))
+        if item['Zone']['Type'] == 'Inactive':
+            g.logger.Debug(2, 'Skipping {} as it is inactive'.format(item['Zone']['Name']))
+            continue
         if  match_reason:
             if not findWholeWord(item['Zone']['Name'])(reason):
                 g.logger.Debug(1,'dropping {} as zones in alarm cause is {}'.format(item['Zone']['Name'], reason))
@@ -121,7 +226,7 @@ def import_zm_zones(mid, reason):
         g.polygons.append({
             'name': item['Zone']['Name'],
             'value': str2tuple(item['Zone']['Coords']),
-            'pattern': g.config.get('object_detection_pattern')
+            'pattern': None
 
         })
 
@@ -129,7 +234,7 @@ def import_zm_zones(mid, reason):
 
 # downloaded ZM image files for future analysis
 def download_files(args):
-    if g.config['wait'] > 0:
+    if int(g.config['wait']) > 0:
         g.logger.Info('Sleeping for {} seconds before downloading'.format(
             g.config['wait']))
         time.sleep(g.config['wait'])
@@ -228,7 +333,7 @@ def download_files(args):
 
 def get_pyzm_config(args):
     g.config['pyzm_overrides'] = {}
-    config_file = ConfigParser(interpolation=None)
+    config_file = ConfigParser(interpolation=None, inline_comment_prefixes='#')
     config_file.read(args.get('config'))
     if config_file.has_option('general', 'pyzm_overrides'):
         pyzm_overrides = config_file.get('general', 'pyzm_overrides')
@@ -261,10 +366,11 @@ def process_config(args, ctx):
 
     def _set_config_val(k, v):
         # internal function to parse all keys
+        #print (f'inside set_config_val with {k}={v}')
         if config_file.has_section(v['section']):
-            val = config_file[v['section']].get(k, v['default'])
+            val = config_file[v['section']].get(k, v.get('default'))
         else:
-            val = v['default']
+            val = v.get('default')
             g.logger.Debug(1,
                 'Section [{}] missing in config file, using key:{} default: {}'
                 .format(v['section'], k, val))
@@ -276,7 +382,7 @@ def process_config(args, ctx):
                     'Secret token found, but no secret file specified')
             if secrets_file.has_option('secrets', val[1:]):
                 vn = secrets_file.get('secrets', val[1:])
-                #g.logger.Debug (1,'Replacing {} with {}'.format(val,vn))
+            #g.logger.Debug (1,'Replacing {} with {}'.format(val,vn))
                 val = vn
             else:
                 raise ValueError(
@@ -292,14 +398,15 @@ def process_config(args, ctx):
 
     # main
     try:
-        config_file = ConfigParser(interpolation=None)
+        config_file = ConfigParser(interpolation=None, inline_comment_prefixes='#')
         config_file.read(args.get('config'))
 
         if config_file.has_option('general', 'secrets'):
             secrets_filename = config_file.get('general', 'secrets')
             g.logger.Debug(1,'secret filename: {}'.format(secrets_filename))
             has_secrets = True
-            secrets_file = ConfigParser(interpolation=None)
+            g.config['secrets'] = secrets_filename
+            secrets_file = ConfigParser(interpolation=None, inline_comment_prefixes='#')
             try:
                 with open(secrets_filename) as f:
                     secrets_file.read_file(f)
@@ -309,32 +416,47 @@ def process_config(args, ctx):
             g.logger.Debug(1,'No secrets file configured')
         # now read config values
 
-        for k, v in g.config_vals.items():
-            #g.logger.Debug (1,'processing {} {}'.format(k,v))
-            if k == 'secrets':
-                continue
+        # first, fill in config with default values
+        for k,v in g.config_vals.items():
+            g.config[k] = v.get('default', None)
 
-            _set_config_val(k, v)
+        # now iterate the file
+        for sec in config_file.sections():
+            if sec.startswith('monitor-'):
+                #g.logger.Debug(4, 'Skipping {} for now'.format(sec))
+                continue
+            if sec == 'secrets':
+                continue
+            for (k, v) in config_file.items(sec):
+                if g.config_vals.get(k):
+                    _set_config_val(k,g.config_vals[k] )
+                else:
+                    #g.logger.Debug(4, 'storing unknown attribute {}={}'.format(k,v))
+                    g.config[k] = v 
+                    #_set_config_val(k,{'section': sec, 'default': None, 'type': 'string'} )
+
         if g.config['allow_self_signed'] == 'yes':
             ctx.check_hostname = False
             ctx.verify_mode = ssl.CERT_NONE
             g.logger.Debug(1,'allowing self-signed certs to work...')
         else:
-            g.logger.Debug(1,'strict SSL cert checking is on...')
+            g.logger.Debug(1,'strict SSL cert checking is on...')           
+
 
         g.polygons = []
         poly_patterns = []
 
-        # Check if we have a custom overrides for this monitor
 
+        # Check if we have a custom overrides for this monitor
+        g.logger.Debug(4,'Now checking for monitor overrides')
         if 'monitorid' in args and args.get('monitorid'):
-            
             sec = 'monitor-{}'.format(args.get('monitorid'))
             if sec in config_file:
                 # we have a specific section for this monitor
                 for item in config_file[sec].items():
                     k = item[0]
                     v = item[1]
+                    g.config[k] = v
 
                     if k.endswith('_zone_detection_pattern'):
                         zone_name = k.split('_zone_detection_pattern')[0]
@@ -344,22 +466,25 @@ def process_config(args, ctx):
 
                     if k in g.config_vals:
                         # This means its a legit config key that needs to be overriden
-                        g.logger.Debug(2,
+                        g.logger.Debug(4,
                             '[{}] overrides key:{} with value:{}'.format(
                                 sec, k, v))
                         g.config[k] = _correct_type(v,
                                                     g.config_vals[k]['type'])
                     else:
                         # This means its a polygon for the monitor
-                        if not g.config['only_triggered_zm_zones'] == 'yes':
-                            try:
-                                g.polygons.append({'name': k, 'value': str2tuple(v),'pattern': g.config.get('object_detection_pattern')})
-                                g.logger.Debug(2,'adding polygon: {} [{}]'.format(k, v ))
-                            except Exception as e:
-                                g.logger.Error('{}={} is either an invalid attribute or a malformed polygon. Error was {}. Ignoring.'.format(k,v,e))
-
+                        if k.startswith(('object_','face_', 'alpr_')):
+                            g.logger.Debug(2,'assuming {} is an ML sequence, adding to config'.format(k))
                         else:
-                            g.logger.Debug (2,'ignoring polygon: {} as only_triggered_zm_zones is true'.format(k))
+                            if not g.config['only_triggered_zm_zones'] == 'yes':
+                                try:
+                                    g.polygons.append({'name': k, 'value': str2tuple(v),'pattern': None})
+                                    g.logger.Debug(2,'adding polygon: {} [{}]'.format(k, v ))
+                                except Exception as e:
+                                    g.logger.Debug(3,'{}={} is not a polygon definition. Error was {}. Ignoring.'.format(k,v,e))
+
+                            else:
+                                g.logger.Debug (2,'ignoring polygon: {} as only_triggered_zm_zones is true'.format(k))
             # now import zones if needed
             # this should be done irrespective of a monitor section
             if g.config['only_triggered_zm_zones'] == 'yes':
@@ -369,7 +494,6 @@ def process_config(args, ctx):
             
             # finally, iterate polygons and put in detection patterns
             for poly in g.polygons:
-
                 for poly_pat in poly_patterns:
                     if poly['name'] == poly_pat['name']:
                         poly['pattern'] = poly_pat['pattern']
@@ -387,20 +511,26 @@ def process_config(args, ctx):
         exit(0)
 
     # Now lets make sure we take care of parameter substitutions {{}}
-
+    g.logger.Debug (4,'Finally, doing parameter substitution')
     p = r'{{(\w+?)}}'
     for gk, gv in g.config.items():
-        if not isinstance(gv, str):
-            continue
-
-        sub_vars = re.findall(p, gv)
-        for sub_var in sub_vars:
-            if g.config[sub_var]:
-
-                g.config[gk] = g.config[gk].replace('{{' + sub_var + '}}',
-                                                    g.config[sub_var])
-                g.logger.Debug(2,'key [{}] is \'{}\' after substitution'.format(
-                    gk, g.config[gk]))
+        #input ('Continue')
+        #print(f"PROCESSING {gk} {gv}")
+        gv = '{}'.format(gv)
+        #if not isinstance(gv, str):
+        #    continue
+        while True:
+            matches = re.findall(p,gv)
+            replaced = False
+            for match_key in matches:
+                if match_key in g.config:
+                    replaced = True
+                    new_val = g.config[gk].replace('{{' + match_key + '}}',str(g.config[match_key]))
+                    #print ('replacing {} with {}'.format(g.config[gk], new_val))
+                    g.config[gk] = new_val
+                    gv = new_val
+            if not replaced:
+                break
 
     # Now munge config if testing args provide
     if args.get('file'):
@@ -414,4 +544,4 @@ def process_config(args, ctx):
         g.config['image_path'] = args.get('output_path')
         g.config['write_debug_image'] = 'yes'
 
-
+