+++ zmeventnotification/zmeventnotification/zmeventnotification.pl
@@ -43,7 +43,7 @@ use Symbol qw(qualify_to_ref);
 use IO::Select;
 
 ####################################
-my $app_version = '6.0.5';
+my $app_version = '6.1.0';
 ####################################
 
 # do this before any log init etc.
@@ -72,7 +72,7 @@ if ( !try_use('JSON') ) {
 }
 
 # debugging only.
-use Data::Dumper;
+#use Data::Dumper;
 
 # ==========================================================================
 #
@@ -123,7 +123,7 @@ use constant {
 
   DEFAULT_CUSTOMIZE_VERBOSE                       => 'no',
   DEFAULT_CUSTOMIZE_EVENT_CHECK_INTERVAL          => 5,
-  DEFAULT_CUSTOMIZE_ES_DEBUG_LEVEL                => 2,
+  DEFAULT_CUSTOMIZE_ES_DEBUG_LEVEL                => 4,
   DEFAULT_CUSTOMIZE_MONITOR_RELOAD_INTERVAL       => 300,
   DEFAULT_CUSTOMIZE_READ_ALARM_CAUSE              => 'no',
   DEFAULT_CUSTOMIZE_TAG_ALARM_EVENT_ID            => 'no',
@@ -140,11 +140,14 @@ use constant {
   DEFAULT_EVENT_END_NOTIFY_ON_HOOK_FAIL      => 'none',
   DEFAULT_EVENT_END_NOTIFY_ON_HOOK_SUCCESS   => 'none',
   DEFAULT_EVENT_END_NOTIFY_IF_START_SUCCESS  => 'yes',
+  DEFAULT_SEND_EVENT_START_NOTIFICATION        => 'yes',
   DEFAULT_SEND_EVENT_END_NOTIFICATION        => 'no',
+
   DEFAULT_USE_ESCONTROL_INTERFACE            => 'no',
   DEFAULT_ESCONTROL_INTERFACE_FILE =>
     '/var/lib/zmeventnotification/misc/escontrol_interface.dat',
   DEFAULT_FCM_DATE_FORMAT => '%I:%M %p, %d-%b',
+  DEFAULT_FCM_ANDROID_PRIORITY=>'default',
   DEFAULT_MAX_FCM_PER_MONTH_PER_TOKEN => 8000
 };
 
@@ -219,6 +222,7 @@ my $api_push_script;
 
 my $token_file;
 my $fcm_date_format;
+my $fcm_android_priority;
 
 my $ssl_enabled;
 my $ssl_cert_file;
@@ -232,6 +236,7 @@ my $read_alarm_cause;
 my $tag_alarm_event_id;
 my $use_custom_notification_sound;
 my $send_event_end_notification;
+my $send_event_start_notification;
 
 my $use_hooks;
 my $event_start_hook;
@@ -329,6 +334,7 @@ GetOptions(
   'version'      => \$version,
   'config=s'     => \$config_file,
   'check-config' => \$check_config,
+  'debug'        => \$console_logs
 );
 
 if ($version) {
@@ -518,6 +524,8 @@ sub loadEsConfigSettings {
 
   $fcm_date_format =
     config_get_val( $config, 'fcm', 'date_format', DEFAULT_FCM_DATE_FORMAT );
+  $fcm_android_priority = 
+    config_get_val( $config, 'fcm', 'android_priority', DEFAULT_FCM_ANDROID_PRIORITY );
 
   $use_api_push =
     config_get_val( $config, 'push', 'use_api_push', DEFAULT_USE_API_PUSH );
@@ -534,7 +542,7 @@ sub loadEsConfigSettings {
   $ssl_cert_file = config_get_val( $config, 'ssl', 'cert' );
   $ssl_key_file  = config_get_val( $config, 'ssl', 'key' );
   $console_logs = config_get_val( $config, 'customize', 'console_logs',
-    DEFAULT_CUSTOMIZE_VERBOSE );
+    DEFAULT_CUSTOMIZE_VERBOSE ) if (!$console_logs);
   $es_debug_level = config_get_val( $config, 'customize', 'es_debug_level',
     DEFAULT_CUSTOMIZE_ES_DEBUG_LEVEL );
   $event_check_interval =
@@ -565,6 +573,10 @@ sub loadEsConfigSettings {
   $send_event_end_notification =
     config_get_val( $config, 'customize', 'send_event_end_notification',
     DEFAULT_SEND_EVENT_END_NOTIFICATION );
+
+  $send_event_start_notification =
+    config_get_val( $config, 'customize', 'send_event_start_notification',
+    DEFAULT_SEND_EVENT_START_NOTIFICATION );
 
   $use_hooks =
     config_get_val( $config, 'customize', 'use_hooks', DEFAULT_USE_HOOKS );
@@ -704,6 +716,8 @@ Use FCM .............................. ${\(yes_or_no($use_fcm))}
 Use FCM V1 APIs....................... ${\(yes_or_no($use_fcmv1))}
 FCM Date Format....................... ${\(value_or_undefined($fcm_date_format))}
 Only show latest FCMv1 message........ ${\(yes_or_no($replace_push_messages))}
+Android FCM push priority............. ${\(value_or_undefined($fcm_android_priority))}
+
 Token file ........................... ${\(value_or_undefined($token_file))}
 
 Use MQTT ............................. ${\(yes_or_no($use_mqtt))}
@@ -727,6 +741,7 @@ ES Debug level.........................${\(value_or_undefined($es_debug_level))}
 Read alarm cause ..................... ${\(yes_or_no($read_alarm_cause))}
 Tag alarm event id ................... ${\(yes_or_no($tag_alarm_event_id))}
 Use custom notification sound ........ ${\(yes_or_no($use_custom_notification_sound))}
+Send event start notification..........${\(yes_or_no($send_event_start_notification))}
 Send event end notification............${\(yes_or_no($send_event_end_notification))}
 Monitor rules JSON file................${\(value_or_undefined($es_rules_file))}
 
@@ -951,19 +966,21 @@ sub printDebug {
   my $str   = shift;
   my $level = shift;
   $level = $es_debug_level if not defined $level;
-  return if $es_debug_level < $level;
   my $now = strftime( '%Y-%m-%d,%H:%M:%S', localtime );
   $str = $prefix . ' ' . $str;
-  print( "CONSOLE DBG-$level:", $now, " ", $str, "\n" ) if $console_logs;
-
-  Debug($str);
+  if ($es_debug_level >= $level) {
+    print( "DBG-$level:", $now, " ", $str, "\n" ) if $console_logs;
+    Debug($str) ;
+  }
+  
+  
 }
 
 sub printInfo {
   my $str = shift;
   my $now = strftime( '%Y-%m-%d,%H:%M:%S', localtime );
   $str = $prefix . ' ' . $str;
-  print( 'CONSOLE INF:', $now, " ", $str, "\n" ) if $console_logs;
+  print( 'INF:', $now, " ", $str, "\n" ) if $console_logs;
 
   Info($str);
 }
@@ -972,7 +989,7 @@ sub printWarning {
   my $str = shift;
   my $now = strftime( '%Y-%m-%d,%H:%M:%S', localtime );
   $str = $prefix . ' ' . $str;
-  print( 'CONSOLE WAR:', $now, " ", $str, "\n" ) if $console_logs;
+  print( 'WAR:', $now, " ", $str, "\n" ) if $console_logs;
   Warning($str);
 }
 
@@ -980,7 +997,7 @@ sub printError {
   my $str = shift;
   my $now = strftime( '%Y-%m-%d,%H:%M:%S', localtime );
   $str = $prefix . ' ' . $str;
-  print( 'CONSOLE ERR:', $now, " ", $str, "\n" ) if $console_logs;
+  print( 'ERR:', $now, " ", $str, "\n" ) if $console_logs;
   Error($str);
 }
 
@@ -1307,19 +1324,22 @@ sub checkNewEvents() {
 
     # use this time to keep token counters updated
     my $update_tokens = 0;
-    open( my $fh, '<', $token_file )
-    || Error( 'Cannot open to update token counts ' . $token_file );
     my %tokens_data;
-    my $hr;
-    my $data = do { local $/ = undef; <$fh> };
-    close($fh);
-    eval { $hr = decode_json($data); };
-    if ($@) {
-      printError("Could not parse token file for token counts: $!");
-    } else {
-      %tokens_data = %$hr;
-      $update_tokens = 1;
+    if ($use_fcm) {
+      open( my $fh, '<', $token_file )
+      || Error( 'Cannot open to update token counts ' . $token_file );
+      my $hr;
+      my $data = do { local $/ = undef; <$fh> };
+      close($fh);
+      eval { $hr = decode_json($data); };
+      if ($@) {
+        printError("Could not parse token file for token counts: $!");
+      } else {
+        %tokens_data = %$hr;
+        $update_tokens = 1;
+      }
     }
+    
 
     # this means we have hit the reload monitor timeframe
     my $len = scalar @active_connections;
@@ -1355,7 +1375,7 @@ sub checkNewEvents() {
       $ndx++;
     }
 
-    if ($update_tokens) {
+    if ($update_tokens && $use_fcm) {
       open( my $fh, '>', $token_file )
       or printError("Error writing tokens file during count update: $!");
       my $json = encode_json( \%tokens_data );
@@ -1620,7 +1640,7 @@ sub validateAuth {
         # perl bcrypt libs can't handle $2b$ or $2y$
         $saved_pass =~ s/^\$2.\$/\$2a\$/;
         my $new_hash = Crypt::Eksblowfish::Bcrypt::bcrypt( $p, $saved_pass );
-        printDebug( "Comparing using bcrypt $new_hash to $saved_pass", 2 );
+        printDebug( "Comparing using bcrypt", 2 );
         return $new_hash eq $saved_pass;
       }
     }
@@ -1882,7 +1902,7 @@ sub sendOverFCMV1 {
   if ( $obj->{platform} eq 'android' ) {
     $message_v2->{android} = {
       icon     => 'ic_stat_notification',
-      priority => 'high'
+      priority => $fcm_android_priority
     };
 
     $message_v2->{android}->{tag} = 'zmninjapush' if ($replace_push_messages);
@@ -1945,31 +1965,19 @@ sub sendOverFCMV1 {
 
   if ( $res->is_success ) {
     $pcnt++;
-    $msg = $res->decoded_content;
     printDebug(
-      'fcmv1: FCM push message returned a 200 with body ' . $res->content, 1 );
-    eval { $json_string = decode_json($msg); };
-    if ($@) {
-
-      Error("fcmv1: Failed decoding sendFCM Response: $@");
-      return;
+      'fcmv1: FCM push message returned a 200 with body ' . $res->decoded_content, 1 );
+  } else {
+    printDebug ('fcmv1: FCM push message error '.$res->decoded_content,1);
+    if ( (index( $res->decoded_content, 'not a valid FCM' ) != -1) ||
+          (index( $res->decoded_content, 'entity was not found') != -1)) {
+      printDebug( 'fcmv1: Removing this token as FCM doesn\'t recognize it',
+        1 );
+      deleteFCMToken( $obj->{token} );
     }
-    if ( $json_string->{failure} eq 1 ) {
-      my $reason = $json_string->{results}[0]->{Error};
-      Error( 'fcmv1: Error sending FCM for token:' . $obj->{token} );
-      Error( 'fcmv1: Error value =' . $reason );
-      if ( index( $reason, 'not a valid FCM' ) != -1 ) {
-        printDebug( 'fcmv1: Removing this token as FCM doesn\'t recognize it',
-          1 );
-        deleteFCMToken( $obj->{token} );
-      }
 
-    }
   }
-  else {
-    printError( 'fcmv1: FCM push message Error:' . $res->status_line );
-  }
-
+    
   # send supplementary event data over websocket, same SSL state issue
   # so use a parent pipe
   if ( $obj->{state} == VALID_CONNECTION && exists $obj->{conn} ) {
@@ -2167,7 +2175,7 @@ sub sendOverFCMLegacy {
   $djson =~ s/pass(word)?=(.*?)($|&)/pass$1=xxx$3/g;
 
   printDebug(
-    "Final JSON being sent is: $djson to token: ..."
+    "legacy: Final JSON being sent is: $djson to token: ..."
       . substr( $obj->{token}, -6 ),
     2
   );
@@ -2458,7 +2466,7 @@ sub checkConnection {
 
   printDebug(
     "After tick: TOTAL: $ac,  ES_CONTROL: $escontrol_conn, FCM+WEB: $fcm_conn, FCM: $fcm_no_conn, WEB: $web_conn, MQTT:$mqtt_conn, invalid WEB: $web_no_conn, PENDING: $pend_conn",
-    3
+    2
   );
 
 }
@@ -2565,7 +2573,10 @@ sub processIncomingMessage {
 
     # This sub type is when a device token is registered
     if ( $json_string->{data}->{type} eq 'token' ) {
-
+      if (!defined($json_string->{data}->{token}) || ($json_string->{data}->{token} eq "")) {
+        printDebug ("Ignoring token command, I got ".encode_json($json_string));
+        return;
+      }
       # a token must have a platform
       if ( !$json_string->{data}->{platform} ) {
         my $str = encode_json(
@@ -2586,6 +2597,10 @@ sub processIncomingMessage {
       my $token_matched = 0;
       my $stored_invocations = undef;
       my $stored_last_sent = undef;
+
+
+      #print Dumper(\@active_connections);
+
       foreach (@active_connections) {
 
         if ( $_->{token} eq $json_string->{data}->{token} ) {
@@ -2601,7 +2616,12 @@ sub processIncomingMessage {
               || $_->{conn}->port() ne $conn->port() )
             )
           {
-            printDebug( 'JOB: token matched but connection did not', 2 );
+            my $existing_token = substr( $_->{token}, -10 );
+            my $new_token = substr( $json_string->{data}->{token}, -10 );
+            my $existing_conn = $_->{conn} ? $_->{conn}->ip().':'.$_->{conn}->port() : 'undefined';
+            my $new_conn = $conn ? $conn->ip().':'.$conn->port() : 'undefined';
+            
+            printDebug( "JOB: new token matched existing token: ($new_token <==> $existing_token) but connection did not ($new_conn <==> $existing_conn)", 2 );
             printDebug(
               'JOB: Duplicate token found: marking ...'
                 . substr( $_->{token}, -10 )
@@ -2630,24 +2650,18 @@ sub processIncomingMessage {
 
             $_->{type}     = FCM;
             $_->{platform} = $json_string->{data}->{platform};
-            if ( exists( $json_string->{data}->{monlist} )
-              && ( $json_string->{data}->{monlist} ne '' ) )
+            if ( isValidMonIntList( $json_string->{data}->{monlist} ))
             {
               $_->{monlist} =
                 $json_string->{data}->{monlist};
             }
-            else {
-              $_->{monlist} = '-1';
-            }
-            if ( exists( $json_string->{data}->{intlist} )
-              && ( $json_string->{data}->{intlist} ne '' ) )
+           
+            if ( isValidMonIntList( $json_string->{data}->{intlist} ))
             {
               $_->{intlist} =
                 $json_string->{data}->{intlist};
             }
-            else {
-              $_->{intlist} = '-1';
-            }
+          
             $_->{pushstate} = $json_string->{data}->{state};
             printDebug(
               'JOB: Storing token ...'
@@ -2678,30 +2692,28 @@ sub processIncomingMessage {
           && ( $_->{conn}->port() eq $conn->port() )
           && ( $_->{token} ne $json_string->{data}->{token} ) )
         {
+          my $existing_token = substr( $_->{token}, -10 );
+          my $new_token = substr( $json_string->{data}->{token}, -10 );
+          my $existing_conn = $_->{conn} ? $_->{conn}->ip().':'.$_->{conn}->port() : 'undefined';
+          my $new_conn = $conn ? $conn->ip().':'.$conn->port() : 'undefined';
+            
           printDebug(
-            'JOB: connection matched but token did not. first registration?',
+            "JOB: connection matched ($new_conn <==> $existing_conn) but token did not ($new_token <==> $existing_token). first registration?",
             2 );
+        
           $_->{type}     = FCM;
           $_->{token}    = $json_string->{data}->{token};
           $_->{platform} = $json_string->{data}->{platform};
-          $_->{monlist}  = $json_string->{data}->{monlist};
-          $_->{intlist}  = $json_string->{data}->{intlist};
-          if ( exists( $json_string->{data}->{monlist} )
-            && ( $json_string->{data}->{monlist} ne '' ) )
-          {
-            $_->{monlist} = $json_string->{data}->{monlist};
-          }
-          else {
-            $_->{monlist} = "-1";
-          }
-          if ( exists( $json_string->{data}->{intlist} )
-            && ( $json_string->{data}->{intlist} ne '' ) )
-          {
-            $_->{intlist} = $json_string->{data}->{intlist};
-          }
-          else {
-            $_->{intlist} = '-1';
-          }
+
+
+          $_->{monlist}  = $json_string->{data}->{monlist}
+            if (isValidMonIntList($json_string->{data}->{monlist}) );
+            
+
+          $_->{intlist}  = $json_string->{data}->{intlist}
+                      if (isValidMonIntList($json_string->{data}->{intlist})) ;
+
+         
           $_->{pushstate} = $json_string->{data}->{state};
           $_->{invocations} = defined ($stored_invocations) ? $stored_invocations:{count=>0, at=>(localtime)[4]};
           #print ("REMOVE applied:". Dumper($_->{invocations}));
@@ -3236,14 +3248,20 @@ sub getInterval {
 
 }
 
+
+sub isValidMonIntList {
+  my $m = shift;
+  printDebug("REMOVE isValid: validating $m",2);
+
+  return defined($m) && ($m ne "-1") && ($m ne "");
+}
 # Checks if the monitor for which
 # an alarm occurred is part of the monitor list
 # for that connection
 sub isInList {
   my $monlist = shift;
   my $mid     = shift;
-
-  #printDebug ("REMOVE fcm: MONLIST=$monlist, MID=$mid");
+  printDebug("REMOVE: looking for $mid inside $monlist",2);
   return 1 if ( $monlist eq "-1" || $monlist eq "" || !$monlist || !defined($monlist) );
 
   my @mids = split( ',', $monlist );
@@ -3308,6 +3326,13 @@ sub sendEvent {
     return;
   }
 
+  if ( ( !$send_event_start_notification ) && ( $event_type eq 'event_start' ) ) {
+    printInfo(
+      'Not sending event start notification as send_event_start_notification is no'
+    );
+    return;
+  }
+
   my $hook = $event_type eq 'event_start' ? $event_start_hook : $event_end_hook;
 
   my $t   = gettimeofday;
@@ -3321,7 +3346,9 @@ sub sendEvent {
 
   if ( $ac->{type} == FCM
     && $ac->{pushstate} ne 'disabled'
-    && $ac->{state} != PENDING_AUTH )
+    && $ac->{state} != PENDING_AUTH 
+    && $ac->{state} != PENDING_DELETE 
+    )
   {
 
     # only send if fcm is an allowed channel