+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Change Log
 
+### 2021-01-02
+- Update Docker baseimage to Focal Alpha (Ubuntu 20.04).
+- Remove SHMEM environment variable and set --shm-size instead.
+- Turn off privileged mode.
+- Update zmNinja Event Notification Server to version 6.1.0.
+
 ### 2020-11-18
 - Add NO_START_ZM environment variable to keep MySql and Zoneminder from starting so the docker stays running and a user can troubleshoot.